Add support for explicitly comparing fields to null
Previously, null was a special case for FieldFilter that automatically forced
it to evaluate to "true", providing an implementation for DefaultFilter.
However, this made it impossible to actually include null in legal filters.
While using null in F# is considered bad practice, it is a very important
concept in Azure Search since all types except for collections are nullable.